Rebase and update did_dht test vector
